Integrating AME-1809 / AME-2156 branch
This merge also includes changes for:
* short reports were always generated when clicking on an actionlink, even
though the report isn't actually displayed.

* Implementation of OPENAM-1630
SAML2 metadata signatures are now compliant with the specification.
Within this change two new realm level option has been introduced:
metadataSigningKey
metadataSigningKeyPass
These will tell OpenAM which private key to use for signing the SAML2
metadata. In case the SAML entity is remote, OpenAM will try to maintain
the original XML signature (if present), otherwise it will be signed with
the configured key.

* Fixing minor bug with configurator in case the system is already upgraded.
The problem was that the request for upgrade.htm resulted in a redirect
which has been automatically followed, so the final response code was 200,
and that confused the configurator.

* Fix for OPENAM-2710

* Fix for OPENAM-2064
Dashboard service has been extracted to a separate LDIF file and the schema
is now available for all supported data store types.